Override identification of gzipped WARCs
Fixes #343

This commit adds logic to the FileIdentifierService allowing it to detect when it has been asked to identify a gzipped WARC file and, when it has, to return the file format information of a non-gzipped WARC file. Requested by product owner to provide more useful information about web archives.